move ResponseType
